store/wizard: extract registration selectors to dedicated module
