+++ src/store/slices/wizard/registration/index.ts
@@ -1,4 +1,5 @@
 export * from './constants';
+export * from './selectors';
 export { initialState as registrationState } from './state';
 export * from './typeguards';
 export * from './types';
